Simplify encryption autoconfiguration. Refactor.
