+++ spring-cloud-config-server/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,7 @@
 # Bootstrap components
 org.springframework.cloud.bootstrap.BootstrapConfiguration=\
 org.springframework.cloud.config.server.bootstrap.ConfigServerBootstrapConfiguration,\
+org.springframework.cloud.config.server.config.TextEncryptionAutoConfiguration,\
 org.springframework.cloud.config.server.config.EncryptionAutoConfiguration
 
 # Environment PostProcessor
@@ -11,6 +12,7 @@ org.springframework.cloud.config.server.bootstrap.ConfigServerBootstrapApplicati
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 org.springframework.cloud.config.server.config.ConfigServerAutoConfiguration,\
 org.springframework.cloud.config.server.config.EncryptionAutoConfiguration,\
+org.springframework.cloud.config.server.config.TextEncryptionAutoConfiguration,\
 org.springframework.cloud.config.server.config.VaultEncryptionAutoConfiguration
 org.springframework.boot.diagnostics.FailureAnalyzer=\
 org.springframework.cloud.config.server.diagnostics.GitUriFailureAnalyzer
